Create Mason.md
I have a very important question (ok maybe not that important but anyways) - Should we write which channel each role is in? Right now, some channels are rather self-explanatory while others can be less obvious. If one happens to be in a channel where the name is not self-explanatory it will say that one is a certain role so idk. Some channel names were going to change as well? I like some funny names tho, so please don't change everything :) Also, when the masons first were introduced it was an idea that they were not allowed to out themselves but since they have been allowed to out themselves the last rounds when they were enabled I wrote that it was allowed. Anyway, masons are probably not going to come into play again so it doesn't really matter.