Update Mason.md
I read what people wrote in the server and agree that there should be no tips in role descriptions. However, I decided to keep that masons are allowed to claim in public channels since that was not evident when the role was introduced. If someone doesn't agree on this, they can, of course, change it :)